group function AVG 제외 구현 및 테스트코드 작성

+++ src/test/java/com/kcb/mqlService/testData/TestDataFactory.java
@@ -31,8 +31,8 @@ public class TestDataFactory {
             Map<String, Object> copied = new HashMap<>();
             for (String eachKey : eachRow.keySet()) {
                 try {
-                    Integer.parseInt(eachRow.get(eachKey));
-                    copied.put(eachKey, Integer.parseInt(eachRow.get(eachKey)));
+                    Double.parseDouble(eachRow.get(eachKey));
+                    copied.put(eachKey, Double.parseDouble(eachRow.get(eachKey)));
                 } catch (Exception e) {
                     copied.put(eachKey, eachRow.get(eachKey));
                 }
